chore: move extensions
