+++ src/Arius.Core/Features/ArchiveCommand/TarBuilder.cs
@@ -1,6 +1,7 @@
 using System.Formats.Tar;
 using Arius.Core.Shared;
 using Arius.Core.Shared.Encryption;
+using Arius.Core.Shared.Extensions;
 
 namespace Arius.Core.Features.ArchiveCommand;
