+++ src/Arius.Core/Shared/Extensions/IEnumerableExtensions.cs
@@ -1,0 +1,9 @@
+namespace Arius.Core.Shared.Extensions;
+
+internal static class IEnumerableExtensions
+{
+    extension<T>(IEnumerable<T> source)
+    {
+        public bool Empty() => !source.Any();
+    }
+}
